requirements.txt
apache-airflow
pandas
kafka-python
requests
fhir.resources